short and unsigned should not be sued for vectors lol. Bounds in position component work fine.Next is checking how to rename and repackage the cursor component into the.... gridmove component... Or not. CursorMove component, and UnitMove component is fine. ALSO I think vector2d is useless and should be removed.

+++ include/vector2D.hpp
@@ -6,11 +6,11 @@
 
 class Vector2D {
     public:
-        short int unsigned x;
-        short int unsigned y;
+        int x;
+        int y;
 
         Vector2D();
-        Vector2D(short unsigned int x, short unsigned int y);
+        Vector2D(int x, int y);
 
         Vector2D & add(const Vector2D & vec);
         Vector2D & subtract(const Vector2D & vec);
